network: Apply context timeout to buffers
When using the network backend when a new buffer is created a new socket is
opened. This socket as well as the associated server side process uses the
default timeout. The intended behavior is that all buffers created for a
context use the timeout specified for the context. Make sure this is the
case by setting the timeout from the context when a new buffer is created.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>